chore: restructure codebase (iteration)
