+++ components/hord-cli/src/core/mod.rs
@@ -1,49 +1,33 @@
 pub mod block;
-pub mod core;
 pub mod inscription;
-pub mod ordinals;
 pub mod pipeline;
+pub mod protocol;
 
 use chainhook_sdk::bitcoincore_rpc::bitcoin::hashes::hex::FromHex;
 use chainhook_sdk::bitcoincore_rpc::bitcoin::Script;
 use chainhook_sdk::bitcoincore_rpc_json::bitcoin::{Txid, Witness};
-use chainhook_sdk::observer::BitcoinConfig;
 use chainhook_sdk::types::{
     BitcoinBlockData, BitcoinNetwork, OrdinalInscriptionCurseType, OrdinalInscriptionRevealData,
-    OrdinalOperation, TransactionIdentifier,
+    OrdinalOperation,
 };
-use crossbeam_channel::bounded;
 use dashmap::DashMap;
 use fxhash::{FxBuildHasher, FxHasher};
-use rand::seq::SliceRandom;
-use rand::thread_rng;
 use rocksdb::DB;
 use rusqlite::Connection;
-use std::collections::{BTreeMap, HashMap, VecDeque};
+use std::collections::BTreeMap;
 use std::hash::BuildHasherDefault;
 use std::ops::Div;
 use std::path::PathBuf;
-use std::sync::mpsc::channel;
-use std::sync::Arc;
-use std::thread::sleep;
-use std::time::Duration;
-use tokio::task::JoinSet;
 
 use chainhook_sdk::{
     bitcoincore_rpc::{Auth, Client, RpcApi},
     utils::Context,
 };
 
-use crate::db::{
-    find_all_inscriptions_in_block, retrieve_satoshi_point_using_lazy_storage_v3, LazyBlock,
-};
-
 use crate::config::{Config, LogConfig};
 
 use chainhook_sdk::indexer::bitcoin::{
-    build_http_client, parse_downloaded_block, standardize_bitcoin_block,
-    try_download_block_bytes_with_retry, BitcoinBlockFullBreakdown,
-    BitcoinTransactionFullBreakdown,
+    standardize_bitcoin_block, BitcoinBlockFullBreakdown, BitcoinTransactionFullBreakdown,
 };
 
 use crate::db::{
@@ -54,7 +38,7 @@ use crate::db::{
 use self::inscription::InscriptionParser;
 use crate::db::{
     insert_transfer_in_locations, remove_entry_from_blocks, remove_entry_from_inscriptions,
-    LazyBlockTransaction, TraversalResult,
+    LazyBlockTransaction,
 };
 use crate::ord::inscription_id::InscriptionId;
