+++ components/hord-cli/src/core/pipeline/processors/mod.rs
@@ -1,0 +1,3 @@
+pub mod inscription_indexing;
+
+pub use inscription_indexing::start_ordinals_number_processor;
