+++ components/hord-cli/src/main.rs
@@ -15,8 +15,8 @@ extern crate serde;
 pub mod archive;
 pub mod cli;
 pub mod config;
+pub mod core;
 pub mod db;
-pub mod hord;
 pub mod ord;
 pub mod scan;
 pub mod service;
